Bugfix: Share component form could not submit
There was an issue where upon pressing the "submit" button in the Share form,
the validation error would always display as "required", preventing the form
form being submitted.

This was due to the form's state being set incorrectly.
This commit fixes that, and removes the unused "name" state and function
arguments.